feat: Add web search functionality
Introduces a new feature that allows the Gemini API to perform web searches. This is controlled by a new `useWebSearch` setting in the `useSettings` store. When enabled, the API will utilize Google Search for relevant queries, enhancing the interactive learning experience by providing up-to-date information. The UI has been updated with a checkbox to toggle this feature.